requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
blas=1.0=mkl
ca-certificates=2018.4.16=0
certifi=2018.4.16=py36_0
cffi=1.11.5=py36h9745a5d_0
cudatoolkit=8.0=3
cycler=0.10.0=py_1
dbus=1.13.0=h3a4f0e9_0
expat=2.2.5=hfc679d8_1
fontconfig=2.13.0=h074f895_2
freetype=2.8.1=0
glib=2.53.6=h5d9569c_2
gst-plugins-base=1.12.4=h33fb286_0
gstreamer=1.12.4=hb53b477_0
icu=58.2=hfc679d8_0
intel-openmp=2018.0.3=0
jpeg=9b=h024ee3a_2
kiwisolver=1.0.1=py36_1
libedit=3.1.20170329=h6b74fdf_2
libffi=3.2.1=hd88cf55_4
libgcc-ng=7.2.0=hdf63c60_3
libgfortran-ng=7.2.0=hdf63c60_3
libiconv=1.15=h470a237_1
libpng=1.6.34=hb9fc6fc_0
libstdcxx-ng=7.2.0=hdf63c60_3
libtiff=4.0.9=he85c1e1_1
libuuid=1.0.3=1
libxcb=1.13=h470a237_1
libxml2=2.9.8=h422b904_2
matplotlib=2.2.2=py36_1
mkl=2018.0.3=1
mkl_fft=1.0.2=py36h651fb7a_0
mkl_random=1.0.1=py36h4414c95_1
ncurses=6.1=hf484d3e_0
ninja=1.8.2=py36h6bb024c_1
numpy=1.14.5=py36h1b885b7_4
numpy-base=1.14.5=py36hdbf6ddf_4
olefile=0.45.1=py36_0
openssl=1.0.2o=0
pcre=8.41=1
pillow=5.2.0=py36_0
pip=10.0.1=py36_0
pthread-stubs=0.4=h470a237_1
pycparser=2.18=py36_1
pyparsing=2.2.0=py_1
pyqt=5.6.0=py36h8210e8a_6
python=3.6.6=hc3d631a_0
python-dateutil=2.7.3=py_0
pytorch=0.4.0=py36_cuda8.0.61_cudnn7.1.2_1
pytz=2018.5=py_0
qt=5.6.2=hd25b39d_14
readline=7.0=ha6073c6_4
setuptools=39.2.0=py36_0
sip=4.18=py36_1
six=1.11.0=py36_1
sqlite=3.24.0=h84994c4_0
tk=8.6.7=hc745277_3
torchvision=0.2.1=py36_1
tornado=5.1=py36_0
wheel=0.31.1=py36_0
xorg-libxau=1.0.8=h470a237_5
xorg-libxdmcp=1.1.2=h470a237_6
xz=5.2.4=h14c3975_4
zlib=1.2.11=ha838bed_2
